move logic operations to lib and better error handling
